move util to ml
